Added CRUD actions for users table
Decided to add userService in a separate services folder for clarity. tRPC setup is still in trpc.ts, considered creating a specific routes folder with userRoutes in it to mimic how we could merge different routers in an _app file but considered that too convoluted for a simple example for now.