Created DeleteUserButton component
We need a button to delete a user in a couple of places so I'm separating it. I've also added functionality to go back to the homepage if we're in the user profile when we delete them. This is done through a callback to be able ;to pass other functionality if needed. I've also updated AddUser to be able to update existing users. I'm not completely happy about the name but I am having trouble thinking of a more fitting name at the moment.